backend/requirements.txt
# 💬 BABA CHAT Backend Dependencies
# =====================================

# 🌐 웹 프레임워크
fastapi==0.104.1              # 현대적인 Python 웹 프레임워크
uvicorn[standard]==0.24.0     # ASGI 서버 (Socket.IO 지원)
python-socketio==5.11.0       # Socket.IO 서버 (비동기 지원)

# 📊 데이터 검증 및 설정
pydantic==2.5.2               # 데이터 검증 및 타입 힌팅
pydantic-settings==2.1.0      # 설정 관리 (환경변수 자동 로드)

# 🔧 유틸리티
python-multipart==0.0.6       # 폼 데이터 처리 (향후 파일 업로드용)
email-validator==2.1.0        # 이메일 검증 (향후 사용자 인증용)

# 🗂️ 정적 파일 서빙
aiofiles==23.2.1               # 비동기 파일 I/O

# 📅 시간 처리
python-dateutil==2.8.2        # 날짜/시간 유틸리티

# 🧪 개발 및 테스트 도구 (선택사항)
# pytest==7.4.3               # 테스트 프레임워크
# pytest-asyncio==0.21.1      # 비동기 테스트 지원
# black==23.11.0               # 코드 포매터
# flake8==6.1.0                # 코드 린터

# 📦 향후 확장용 (주석 처리됨)
# redis==5.0.1                 # Redis 클라이언트 (세션 저장용)
# sqlalchemy==2.0.23           # ORM (데이터베이스 연동용)
# alembic==1.13.1              # 데이터베이스 마이그레이션
# passlib[bcrypt]==1.7.4       # 비밀번호 해싱
# python-jose[cryptography]==3.3.0  # JWT 토큰 처리
# slowapi==0.1.9               # 레이트 리미팅